+++ src/CG.Business/Strategies/ApplicationBuilderExtensions.cs
@@ -6,18 +6,20 @@ using CG.DataAnnotations;
 using CG.Reflection;
 using CG.Validations;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using System;
 using System.IO;
 using System.Linq;
 using System.Reflection;
 
-namespace Microsoft.Extensions.DependencyInjection
+namespace Microsoft.AspNetCore.Builder
 {
     /// <summary>
-    /// This class contains extension methods related to the <see cref="IServiceCollection"/>
+    /// This class contains extension methods related to the <see cref="IApplicationBuilder"/>
     /// type, for registering types related to strategies.
     /// </summary>
-    public static partial class StrategiesServiceCollectionExtensions
+    public static partial class ApplicationBuilderExtensions
     {
         // *******************************************************************
         // Public methods.
@@ -28,38 +30,38 @@ namespace Microsoft.Extensions.DependencyInjection
         /// <summary>
         /// This method reads the configuration for a <see cref="LoaderOptions"/>
         /// compatible section. It then uses that information to dynamically locate,
-        /// and then invoke, an extension method that registers concrete options 
-        /// and strategy types. 
+        /// and then invoke, an extension method that wires up any startup logic
+        /// required to run the configured strategy type(s).
         /// </summary>
-        /// <param name="serviceCollection">The service collection to use 
-        /// for the operation.</param>
-        /// <param name="configuration">The configuration to use for the 
-        /// operation.</param>
-        /// <param name="serviceLifetime">The service lifetime to use for the operation.</param>
-        /// <param name="assemblyWhiteList">An optional white list for filtering
-        /// the list of assemblies that are searched during this operation.</param>
+        /// <param name="applicationBuilder">The application builder to use for 
+        /// the operation.</param>
+        /// <param name="hostEnvironment">The hosting environment to use for the operation.</param>
+        /// <param name="configuration">The configuration section to use for the operation.</param>
         /// <param name="assemblyBlackList">An optional black list for filtering
         /// the list of assemblies that are searched during this operation.</param>
-        /// <returns>the value of the <paramref name="serviceCollection"/>
+        /// <param name="assemblyWhiteList">An optional white list for filtering
+        /// the list of assemblies that are searched during this operation.</param>
+        /// <returns>The value of the <paramref name="applicationBuilder"/>
         /// parameter, for chaining calls together.</returns>
-        /// <exception cref="ArgumentException">This exception is thrown whenever
-        /// one ore more of the parameters is missing, or invalid.</exception>
+        /// <exception cref="ArgumentException">This exception is thrown whenever one
+        /// or more arguments are invalid, or missing.</exception>
         /// <exception cref="ConfigurationException">This exception is thrown whenever
         /// the method failes to locate appropriate configuration sections and/or settings
         /// at runtime.</exception>
         /// <exception cref="BusinessException">This exception is thrown whenever the 
         /// method failes to locate an appropriate extension method to call, or when that
         /// call fails.</exception>
-        public static IServiceCollection AddStrategies(
-            this IServiceCollection serviceCollection,
+        public static IApplicationBuilder UseStrategies(
+            this IApplicationBuilder applicationBuilder,
+            IHostEnvironment hostEnvironment,
             IConfiguration configuration,
-            ServiceLifetime serviceLifetime = ServiceLifetime.Scoped,
             string assemblyWhiteList = "",
             string assemblyBlackList = "Microsoft*, System*, mscorlib, netstandard"
             )
         {
             // Validate the parameters before attempting to use them.
-            Guard.Instance().ThrowIfNull(serviceCollection, nameof(serviceCollection))
+            Guard.Instance().ThrowIfNull(applicationBuilder, nameof(applicationBuilder))
+                .ThrowIfNull(hostEnvironment, nameof(hostEnvironment))
                 .ThrowIfNull(configuration, nameof(configuration));
 
             // Create the loader options.
@@ -75,7 +77,7 @@ namespace Microsoft.Extensions.DependencyInjection
                 throw new ConfigurationException(
                     message: string.Format(
                         Resources.InvalidLoaderSection,
-                        nameof(AddStrategies),
+                        nameof(UseStrategies),
                         configuration.GetPath()
                         )
                     );
@@ -91,7 +93,7 @@ namespace Microsoft.Extensions.DependencyInjection
                 throw new ConfigurationException(
                     message: string.Format(
                         Resources.EmptyStrategyName,
-                        nameof(AddStrategies)
+                        nameof(UseStrategies)
                         )
                     );
             }
@@ -136,7 +138,7 @@ namespace Microsoft.Extensions.DependencyInjection
             catch (Exception ex)
             {
                 // Provide better context for the error.
-                throw new ConfigurationException(
+                throw new BusinessException(
                     message: string.Format(
                         Resources.NoLoadAssembly,
                         loaderOptions.AssemblyNameOrPath
@@ -146,13 +148,13 @@ namespace Microsoft.Extensions.DependencyInjection
             }
 
             // Format the name of a target extension method.
-            var methodName = $"Add{strategyName}Strategies";
+            var methodName = $"Use{strategyName}Strategies";
 
             // Look for specified extension method.
             var methods = AppDomain.CurrentDomain.ExtensionMethods(
-                typeof(IServiceCollection),
+                typeof(IApplicationBuilder),
                 methodName,
-                new Type[] { typeof(IConfiguration), typeof(ServiceLifetime) },
+                new Type[] { typeof(string) },
                 assemblyWhiteList,
                 assemblyBlackList
                 );
@@ -160,23 +162,14 @@ namespace Microsoft.Extensions.DependencyInjection
             // Did we find it?
             if (methods.Any())
             {
-                // Drill down to the right configuration sub-section.
-                var subSection = configuration.GetSection(
-                    strategyName
-                    );
-
                 // We'll use the first matching method.
                 var method = methods.First();
 
                 // Invoke the extension method.
                 method.Invoke(
                     null,
-                    new object[]
-                    {
-                    serviceCollection,
-                    subSection,
-                    serviceLifetime
-                    });
+                    new object[] { applicationBuilder, hostEnvironment, configuration }
+                    );
             }
             else
             {
@@ -184,18 +177,18 @@ namespace Microsoft.Extensions.DependencyInjection
                 //   extension method!
 
                 // Panic!
-                throw new ConfigurationException(
+                throw new BusinessException(
                     message: string.Format(
                         Resources.MethodNotFound,
-                        nameof(AddStrategies),
+                        nameof(UseStrategies),
                         methodName,
-                        $"{nameof(IServiceCollection)},{nameof(IConfiguration)}, {nameof(ServiceLifetime)}"
+                        $"{nameof(IApplicationBuilder)},{nameof(IHostEnvironment)},{nameof(IConfiguration)}"
                         )
                     );
             }
 
-            // Return the service collection.
-            return serviceCollection;
+            // Return the application builder.
+            return applicationBuilder;
         }
 
         #endregion